Pluralize endpoints which return multiple values
These endpoints return more than one result, so it makes sense to have
their names be plural to communicate that.

This removes a previous parity between the keys (or "dimensions")
allowed as query parameters to the `/reach_counts` endpoint. These
allowed values can be found in the `docs/reach_count.md` file.